+++ admin_ui/src/components/KeySearch.vue
@@ -29,11 +29,12 @@
 </template>
 
 <script lang="ts">
+import { defineComponent } from "vue"
 import KeySearchModal from "./KeySearchModal.vue"
 
 let persistedData = undefined
 
-export default {
+export default defineComponent({
     props: {
         fieldName: String,
         tableName: String,
@@ -79,5 +80,5 @@ export default {
         this.selectedValue = persistedData
         this.hiddenSelectedValue = this.$route.query[this.fieldName]
     }
-}
+})
 </script>